Hotfix user rights checking

+++ src/wishes/wishes.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, ForbiddenException } from '@nestjs/common';
 import { Repository, FindManyOptions } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 
@@ -35,12 +35,41 @@ export class WishesService {
     });
   }
 
-  async update(id: number, wishNewData: UpdateWishDto) {
+  async update(id: number, wishNewData: UpdateWishDto, user: User) {
+    const editedWish = await this.wishRepository.findOne({
+      where: { id },
+      relations: {
+        owner: true,
+        offers: true,
+      },
+    });
+    if (
+      user.id !== editedWish.owner.id &&
+      !wishNewData.hasOwnProperty('raised')
+    ) {
+      throw new ForbiddenException('Вы не можете редактировать чужие подарки');
+    }
+    if (wishNewData.price && editedWish.raised > 0) {
+      throw new ForbiddenException(
+        'Вы не можете изменять стоимость подарка, если уже есть желающие скинуться',
+      );
+    }
     await this.wishRepository.update(id, wishNewData);
     return {};
   }
 
-  remove(id: number) {
+  async remove(id: number, user: User) {
+    const deletingWish = await this.wishRepository.findOne({
+      where: { id },
+      relations: {
+        owner: true,
+        offers: true,
+      },
+    });
+    if (user.id !== deletingWish.owner.id) {
+      throw new ForbiddenException('Вы не можете удалять чужие подарки');
+    }
+
     return this.wishRepository.delete({ id });
   }
 
@@ -64,7 +93,32 @@ export class WishesService {
     });
   }
 
-  async copyWish(wish: Wish) {
+  async copyWish(wish: Wish, user: User) {
+    if (user.id === wish.owner.id) {
+      throw new ForbiddenException('Вы не можете копировать свой подарок');
+    }
+
+    const arrayForSameWishes = await this.wishRepository.find({
+      where: [{ name: wish.name }],
+      relations: {
+        owner: true,
+      },
+    });
+
+    let hasUserCopied = false;
+
+    arrayForSameWishes.forEach((item) => {
+      if (+item.owner.id === +user.id) {
+        hasUserCopied = true;
+      }
+    });
+
+    if (hasUserCopied) {
+      throw new ForbiddenException(
+        'Вы уже создали у себя или скопировали себе такой же подарок',
+      );
+    }
+
     const dataWish = {
       name: wish.name,
       image: wish.image,
